requirements.txt
# Core ML and data processing
torch
numpy>=1.24.0
pandas>=2.0.0
scikit-learn

# Technical analysis
ta-lib
pandas-ta

# Data download and visualization
yfinance
matplotlib
seaborn

# Utilities
python-dotenv
tqdm

# Binance API
binance-connector>=3.5.0
python-binance>=1.0.19
aiohttp>=3.8.0